Refactor JDK externalizers into wildfly-clustering-marshalling-spi.
