+++ clustering/marshalling/spi/src/main/java/org/wildfly/clustering/marshalling/spi/IndexExternalizer.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.clustering.marshalling.jboss;
+package org.wildfly.clustering.marshalling.spi;
 
 import java.io.DataInput;
 import java.io.DataOutput;
