+++ clustering/marshalling/spi/src/main/java/org/wildfly/clustering/marshalling/spi/io/ExternalizableExternalizer.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.wildfly.clustering.marshalling.jboss;
+package org.wildfly.clustering.marshalling.spi.io;
 
 import java.io.Externalizable;
 import java.io.IOException;
