+++ clustering/marshalling/spi/src/main/java/org/wildfly/clustering/marshalling/spi/util/CollectionExternalizer.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.clustering.marshalling.jboss;
+package org.wildfly.clustering.marshalling.spi.util;
 
 import java.io.IOException;
 import java.io.ObjectInput;
@@ -37,6 +37,7 @@ import java.util.function.IntFunction;
 
 import org.kohsuke.MetaInfServices;
 import org.wildfly.clustering.marshalling.Externalizer;
+import org.wildfly.clustering.marshalling.spi.IndexExternalizer;
 
 /**
  * Externalizers for non-concurrent implementations of {@link Collection}.
