+++ clustering/marshalling/spi/src/main/java/org/wildfly/clustering/marshalling/spi/util/OptionalExternalizer.java
@@ -1,6 +1,6 @@
 /*
  * JBoss, Home of Professional Open Source.
- * Copyright 2015, Red Hat, Inc., and individual contributors
+ * Copyright 2016, Red Hat, Inc., and individual contributors
  * as indicated by the @author tags. See the copyright.txt file in the
  * distribution for a full listing of individual contributors.
  *
@@ -20,26 +20,37 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.clustering.marshalling.jboss;
+package org.wildfly.clustering.marshalling.spi.util;
 
 import java.io.IOException;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.concurrent.CopyOnWriteArraySet;
+import java.io.ObjectInput;
+import java.io.ObjectOutput;
+import java.util.Optional;
 
-import org.junit.Test;
+import org.kohsuke.MetaInfServices;
+import org.wildfly.clustering.marshalling.Externalizer;
 
 /**
- * Unit test for {@link CopyOnWriteCollectionExternalizer} externalizers.
+ * {@link Externalizer} for {@link Optional} values.
  * @author Paul Ferraro
  */
-public class CopyOnWriteCollectionExternalizerTestCase {
+@MetaInfServices(Externalizer.class)
+public class OptionalExternalizer implements Externalizer<Optional<Object>> {
 
-    @Test
-    public void test() throws ClassNotFoundException, IOException {
-        Collection<Object> basis = Arrays.<Object>asList(1, 2, 3, 4, 5);
-        CollectionExternalizerTestCase.test(new CopyOnWriteCollectionExternalizer.CopyOnWriteArrayListExternalizer(), new CopyOnWriteArrayList<>(basis));
-        CollectionExternalizerTestCase.test(new CopyOnWriteCollectionExternalizer.CopyOnWriteArraySetExternalizer(), new CopyOnWriteArraySet<>(basis));
+    @Override
+    public void writeObject(ObjectOutput output, Optional<Object> optional) throws IOException {
+        output.writeObject(optional.orElse(null));
+    }
+
+    @Override
+    public Optional<Object> readObject(ObjectInput input) throws IOException, ClassNotFoundException {
+        return Optional.ofNullable(input.readObject());
+    }
+
+    @SuppressWarnings({ "rawtypes", "unchecked" })
+    @Override
+    public Class<? extends Optional<Object>> getTargetClass() {
+        Class targetClass = Optional.class;
+        return targetClass;
     }
 }
