+++ clustering/marshalling/spi/src/main/java/org/wildfly/clustering/marshalling/spi/util/UUIDExternalizer.java
@@ -1,6 +1,6 @@
 /*
  * JBoss, Home of Professional Open Source.
- * Copyright 2015, Red Hat, Inc., and individual contributors
+ * Copyright 2012, Red Hat, Inc., and individual contributors
  * as indicated by the @author tags. See the copyright.txt file in the
  * distribution for a full listing of individual contributors.
  *
@@ -19,28 +19,36 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-
-package org.wildfly.clustering.marshalling.jboss;
+package org.wildfly.clustering.marshalling.spi.util;
 
 import java.io.IOException;
-import java.util.Map;
-import java.util.TreeMap;
-import java.util.concurrent.ConcurrentSkipListMap;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
+import java.io.ObjectInput;
+import java.io.ObjectOutput;
+import java.util.UUID;
 
-import org.junit.Test;
+import org.kohsuke.MetaInfServices;
+import org.wildfly.clustering.marshalling.Externalizer;
 
 /**
- * Unit test for {@link SortedSetExternalizer} externalizers.
+ * {@link Externalizer} for {@link UUID} instances.
  * @author Paul Ferraro
  */
-public class SortedMapExternalizerTestCase {
+@MetaInfServices(Externalizer.class)
+public class UUIDExternalizer implements Externalizer<UUID> {
 
-    @Test
-    public void test() throws ClassNotFoundException, IOException {
-        Map<Object, Object> basis = Stream.of(1, 2, 3, 4, 5).collect(Collectors.<Integer, Object, Object>toMap(i -> i, i -> Integer.toString(i)));
-        ExternalizerTestUtil.test(new SortedMapExternalizer.ConcurrentSkipListMapExternalizer(), new ConcurrentSkipListMap<>(basis));
-        ExternalizerTestUtil.test(new SortedMapExternalizer.TreeMapExternalizer(), new TreeMap<>(basis));
+    @Override
+    public void writeObject(ObjectOutput output, UUID uuid) throws IOException {
+        output.writeLong(uuid.getMostSignificantBits());
+        output.writeLong(uuid.getLeastSignificantBits());
+    }
+
+    @Override
+    public UUID readObject(ObjectInput input) throws IOException {
+        return new UUID(input.readLong(), input.readLong());
+    }
+
+    @Override
+    public Class<UUID> getTargetClass() {
+        return UUID.class;
     }
 }
