+++ clustering/marshalling/spi/src/test/java/org/wildfly/clustering/marshalling/spi/util/SortedMapExternalizerTestCase.java
@@ -1,6 +1,6 @@
 /*
  * JBoss, Home of Professional Open Source.
- * Copyright 2013, Red Hat, Inc., and individual contributors
+ * Copyright 2015, Red Hat, Inc., and individual contributors
  * as indicated by the @author tags. See the copyright.txt file in the
  * distribution for a full listing of individual contributors.
  *
@@ -19,30 +19,29 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.wildfly.clustering.web.infinispan.session.fine;
+
+package org.wildfly.clustering.marshalling.spi.util;
 
 import java.io.IOException;
-import java.io.ObjectOutput;
+import java.util.Map;
+import java.util.TreeMap;
+import java.util.concurrent.ConcurrentSkipListMap;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
-import org.kohsuke.MetaInfServices;
-import org.wildfly.clustering.marshalling.Externalizer;
-import org.wildfly.clustering.marshalling.spi.IndexExternalizer;
-import org.wildfly.clustering.web.infinispan.SessionKeyExternalizer;
+import org.junit.Test;
+import org.wildfly.clustering.marshalling.spi.ExternalizerTestUtil;
 
 /**
- * Externalizer for a {@link SessionAttributeKey}.
+ * Unit test for {@link SortedMapExternalizer} externalizers.
  * @author Paul Ferraro
  */
-@MetaInfServices(Externalizer.class)
-public class SessionAttributeKeyExternalizer extends SessionKeyExternalizer<SessionAttributeKey> {
+public class SortedMapExternalizerTestCase {
 
-    public SessionAttributeKeyExternalizer() {
-        super(SessionAttributeKey.class, (id, input) -> new SessionAttributeKey(id, IndexExternalizer.VARIABLE.readData(input)));
-    }
-
-    @Override
-    public void writeObject(ObjectOutput output, SessionAttributeKey key) throws IOException {
-        super.writeObject(output, key);
-        IndexExternalizer.VARIABLE.writeData(output, key.getAttributeId());
+    @Test
+    public void test() throws ClassNotFoundException, IOException {
+        Map<Object, Object> basis = Stream.of(1, 2, 3, 4, 5).collect(Collectors.<Integer, Object, Object>toMap(i -> i, i -> Integer.toString(i)));
+        ExternalizerTestUtil.test(new SortedMapExternalizer.ConcurrentSkipListMapExternalizer(), new ConcurrentSkipListMap<>(basis));
+        ExternalizerTestUtil.test(new SortedMapExternalizer.TreeMapExternalizer(), new TreeMap<>(basis));
     }
 }
