+++ feature-pack/src/main/resources/modules/system/layers/base/org/wildfly/clustering/marshalling/spi/main/module.xml
@@ -29,5 +29,7 @@
     </resources>
 
     <dependencies>
+        <module name="org.wildfly.clustering.marshalling.api"/>
+        <module name="org.wildfly.security.elytron"/>
     </dependencies>
 </module>
