+++ feature-pack/src/main/resources/modules/system/layers/base/org/wildfly/clustering/server/main/module.xml
@@ -55,7 +55,7 @@
         <module name="org.wildfly.clustering.jgroups.spi"/>
         <module name="org.wildfly.clustering.marshalling.api"/>
         <module name="org.wildfly.clustering.marshalling.jboss" services="import"/>
-        <module name="org.wildfly.clustering.marshalling.spi"/>
+        <module name="org.wildfly.clustering.marshalling.spi" services="import"/>
         <module name="org.wildfly.clustering.service"/>
         <module name="org.wildfly.clustering.singleton"/>
         <module name="org.wildfly.clustering.spi"/>
